+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.24
+
+* `measure.time_remaining` will not return negative values